atapiscsi didnt call scsi_done when the xfer had SCSI_POLL set. this is Bad.
this sort of worked before because the midlayer would sleep on the
ITSDONE flag in the scsi_xfer, which atapiscsi does set. however, with
the introduction of scsi_xs_sync, it now sleeps on a condition that is
only cleared if the adapter properly and correctly calls scsi_done when
it is in fact done with the xfer.

sorry guys.

ok krw@