remove horrible abuse of kthread_create_deferred(9).  this was being used to "do things later" and was potentially sleeping in swapper context.  the latter is really bad, because it might never wake up.
instead, move the things that usb_first_explore() was doing to the
first run of the explore task.  since tasks are run in a kthread,
they are in process context and ok to sleep.

fixes boot hang noticed by jsg@, which mikeb@ debugged and found
usb_first_explore() to be sleeping swapper context