From an email to tech@:
Bit difficult to explain this one. If you have a file open that is 3
or 4 times longer than the length of the viewable window and are at
the bottom of the buffer then scroll up to the top using M-v, your
cursor should remain at the bottom left of the window once you reach
the top of the buffer.

mg behaves like this since the last couple of scrolling diffs I've
committed. However, and as an example, if you start with the window
viewing the top of a buffer and press C-v twice, then C-p four times
then M-v twice to scroll back to the top, you will find your cursor is
NOT at the bottom left of the window. tut tut. This diff makes the
cursor stay at the bottom left, like emacs.