stupid code bug.
instead of initting the ccbs sgl dma virtual address, i was assigning it
to the sense memories dva. this means we wouldnt be getting sense data
where we thought it was, and passing NULL (does NULL mean anything over
a pci bus?) to the hardware when we have long sgl chains.

hopefully this fixes the NMIs ive been getting.