Fix a race condition which would cause segfault due to the kernel sending less (or more) data than expected.
We do a sysctl to know how much data should be read, and then we try to
read that amount, but there is a window between this two calls that
things can change, this makes sure we have an "atomic view" of data.

From Patrick Wildt, tested with over 7000 SAs, thanks.

ok deraadt