Stop providing our own EXPAND_BUILTIN_SAVEREGS(). Instead, do the necessary RTL generation in SETUP_INCOMING_VARARGS(...,0).
This means we no longer need to explicitely allocate room on the stack for
the saved register area (which we did in EXPAND_BUILTIN_SAVEREGS). We were
returning the correct size in SETUP_INCOMING_VARARGS's ``pretend_size''
argument anyway, causing this space to be allocated twice.

This change should also allow for the stack protector to be reenabled (after
more testing).