Add a luna88k-specific function to initialize the instruction cmmu SAPR register.
On luna88k, secondary processors are not disabled while the kernel is
initializing.  They are running an infinite loop in
locore.S:secondary_init on non-MULTIPROCESSOR kernel.  Then, after
initializing the CMMUs tied to the currently-running processor, we
turn on the instruction cache of other processors to make them
happier.

As a bonus, on non-MULTIPROCESSOR kernel, the power switch LED of
luna88k is blinking green and orange again.  According to the hardware
manual, it indicates processors' cache hit status.

suggested and ok miod@