Unattended installation using DHCP and a response file
For a completely unattended installation bsd.rd has to be netbooted,
a DHCP server must be running and provide "next-server", which will be
used to fetch "http://<next-server>/install.conf".  The format of the
response file is a list of "<key> = <value>" pairs where <key> is a
substring of the interactive question (case-insensitive) and <value> is
what would be entered interactively.

Minimal response file example:

  system hostname = openbsd
  password for root account = <...>
  network interfaces = re0
  IPv4 address for re0 = dhcp
  server? = <...>

This is a starting point, it still a bit rough.

ok krw@, many improvements by halex@