rework mpii to use tasks instead of scsi_tasks for sas event handling. scsi_tasks are still backed by workq, and i dont like having to allocate storage to defer work.
sas events that cause hotplug events are queued in the softc to be
handled by a single task that loops on the queue for work. think
levelled interrupts. this has a side effect that interrupts are
necessary to enumerate devices for attachment.

i wrote this toward the end of last year, but thought it best to
wait for 5.5 unlock.